Use ghc-lib, compiling with ghc-8.8.3.
