+++ defaults.dhall
@@ -1,0 +1,15 @@
+{ version =
+    "0.5"
+, author =
+    "Christiaan Baaij"
+, maintainer =
+    "christiaan.baaij@gmail.com"
+, copyright =
+    ''
+    Copyright © 2015-2016, University of Twente,
+                                2017-2018, QBayLogic''
+, tested-with =
+    "GHC == 7.10.3, GHC == 8.0.2, GHC == 8.2.2, GHC == 8.4.4, GHC == 8.6.5, GHC == 8.8.2, GHC == 8.10.1"
+, extra-source-files =
+    [ "README.md", "CHANGELOG.md", "defaults.dhall", "package.dhall" ]
+}
