+++ ghc-tcplugins-extra.cabal
@@ -1,47 +1,61 @@
-name:                ghc-tcplugins-extra
-version:             0.4.2
-synopsis:            Utilities for writing GHC type-checker plugins
-description:         Utilities for writing GHC type-checker plugins, such as
-                     creating constraints, with a stable API covering multiple
-                     GHC releases.
-homepage:            http://github.com/clash-lang/ghc-tcplugins-extra
-bug-reports:         http://github.com/clash-lang/ghc-tcplugins-extra/issues
-license:             BSD2
-license-file:        LICENSE
-author:              Christiaan Baaij
-maintainer:          christiaan.baaij@gmail.com
-copyright:           Copyright © 2015-2016, University of Twente,
-                                 2017-2018, QBayLogic
-category:            Type System
-build-type:          Simple
-extra-source-files:  README.md
-                     CHANGELOG.md
-cabal-version:       >=1.10
-tested-with:         GHC == 7.10.3, GHC == 8.0.2, GHC == 8.2.2, GHC == 8.4.4,
-                     GHC == 8.6.5, GHC == 8.8.4, GHC == 8.10.4, GHC == 9.0.2,
-                     GHC == 9.2.1
+cabal-version: 1.12
+
+-- This file has been generated from package.dhall by hpack version 0.31.0.
+--
+-- see: https://github.com/sol/hpack
+--
+-- hash: ce175bdd799abe9b42f478093cc658262f8e3d8cc8744f903081fe653072e690
+
+name:           ghc-tcplugins-extra
+version:        0.4.2
+synopsis:       Utilities for writing GHC type-checker plugins
+description:    Utilities for writing GHC type-checker plugins, such as
+                creating constraints, with a stable API covering multiple
+                GHC releases.
+category:       Type System
+homepage:       https://github.com/clash-lang/ghc-tcplugins-extra#readme
+bug-reports:    https://github.com/clash-lang/ghc-tcplugins-extra/issues
+author:         Christiaan Baaij
+maintainer:     christiaan.baaij@gmail.com
+copyright:      Copyright © 2015-2016, University of Twente,
+                            2017-2018, QBayLogic
+license:        BSD2
+license-file:   LICENSE
+tested-with:    GHC == 7.10.3, GHC == 8.0.2, GHC == 8.2.2, GHC == 8.4.4, GHC == 8.6.5, GHC == 8.8.2, GHC == 8.10.1
+build-type:     Simple
+extra-source-files:
+  README.md
+  CHANGELOG.md
+  defaults.dhall
+  package.dhall
 
 source-repository head
   type: git
-  location: https://github.com/clash-lang/ghc-tcplugins-extra.git
+  location: https://github.com/clash-lang/ghc-tcplugins-extra
 
 flag deverror
-  description:
-    Enables `-Werror` for development mode and TravisCI
-  default: False
+  description: Enables `-Werror` for development mode and TravisCI
   manual: True
+  default: False
 
 library
-  exposed-modules:     GHC.TcPluginM.Extra
-  build-depends:       base >=4.8  && <5,
-                       ghc  >=7.10 && <9.4
-  hs-source-dirs:      src
-  default-language:    Haskell2010
-  other-extensions:    CPP
-                       LambdaCase
-                       RecordWildCards
-                       PatternSynonyms
-  if flag(deverror)
-    ghc-options:         -Wall -Werror
+  exposed-modules:
+    GHC.TcPluginM.Extra
+  other-modules:
+    GhcApi
+  hs-source-dirs:
+    src
+  other-extensions: CPP LambdaCase RecordWildCards PatternSynonyms
+  build-depends:
+    base >=4.8 && <5
+  if impl(ghc > 8.4.4)
+    hs-source-dirs:
+      src-ghc-lib
+    build-depends:
+      ghc-lib
   else
-    ghc-options:         -Wall
+    hs-source-dirs:
+      src-ghc
+    build-depends:
+      ghc >=7.10 && <8.12
+  default-language: Haskell2010
